Reorganize and simplify 1. Simplify GitUtils functions; one function for parsing logs and determining merge status 2. Move parsing and io logic into own modules 3. Organize module hierarchy
